refactor: split hermes-config schema into expandable entities/hermes-config/ dir
