+++ apps/dashboard/src/entities/index.ts
@@ -3,3 +3,4 @@ export * from "./agent";
 export * from "./template";
 export * from "./hermeum-config";
 export * from "./shared-env-set";
+export * from "./hermes-config";
